refactor: rename 'json' package to 'jsonutils'
