+++ internal/handlers/handlers.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/vladyslavpavlenko/genesis-api-project/pkg/json"
+	"github.com/vladyslavpavlenko/genesis-api-project/pkg/jsonutils"
 )
 
 // rateUpdate holds the exchange rateapi update data.
@@ -20,12 +20,12 @@ func (m *Repository) GetRate(w http.ResponseWriter, r *http.Request) {
 	// Perform the fetching operation
 	price, err := m.Services.Fetcher.Fetch(r.Context(), "USD", "UAH")
 	if err != nil {
-		_ = json.ErrorJSON(w, fmt.Errorf("error fetching rate update: %w", err), http.StatusServiceUnavailable)
+		_ = jsonutils.ErrorJSON(w, fmt.Errorf("error fetching rate update: %w", err), http.StatusServiceUnavailable)
 		return
 	}
 
 	// AddSubscription a response
-	payload := json.Response{
+	payload := jsonutils.Response{
 		Error: false,
 		Data: rateUpdate{
 			BaseCode:   "USD",
@@ -35,12 +35,12 @@ func (m *Repository) GetRate(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Send the response back
-	_ = json.WriteJSON(w, http.StatusOK, payload)
+	_ = jsonutils.WriteJSON(w, http.StatusOK, payload)
 }
 
 // subscriptionBody is the email subscription request body structure.
 type subscriptionBody struct {
-	Email string `json:"email"`
+	Email string `jsonutils:"email"`
 }
 
 // Subscribe handles the `/subscribe` request.
@@ -50,31 +50,31 @@ func (m *Repository) Subscribe(w http.ResponseWriter, r *http.Request) {
 
 	err := r.ParseMultipartForm(10 << 20)
 	if err != nil {
-		_ = json.ErrorJSON(w, errors.New("failed to parse form"))
+		_ = jsonutils.ErrorJSON(w, errors.New("failed to parse form"))
 		return
 	}
 
 	body.Email = r.FormValue("email")
 	if body.Email == "" {
-		_ = json.ErrorJSON(w, errors.New("email is required"))
+		_ = jsonutils.ErrorJSON(w, errors.New("email is required"))
 		return
 	}
 
 	// Perform the subscription operation
 	code, err := m.SubscribeUser(body.Email)
 	if err != nil {
-		_ = json.ErrorJSON(w, err, code)
+		_ = jsonutils.ErrorJSON(w, err, code)
 		return
 	}
 
 	// AddSubscription a response
-	payload := json.Response{
+	payload := jsonutils.Response{
 		Error:   false,
 		Message: "subscribed",
 	}
 
 	// Send the response back
-	_ = json.WriteJSON(w, http.StatusOK, payload)
+	_ = jsonutils.WriteJSON(w, http.StatusOK, payload)
 }
 
 // SendEmails handles the `/sendEmails` request.
@@ -82,16 +82,16 @@ func (m *Repository) SendEmails(w http.ResponseWriter, _ *http.Request) {
 	// Perform the mailing operation
 	err := m.NotifySubscribers()
 	if err != nil {
-		_ = json.ErrorJSON(w, err, http.StatusInternalServerError)
+		_ = jsonutils.ErrorJSON(w, err, http.StatusInternalServerError)
 		return
 	}
 
 	// AddSubscription a response
-	payload := json.Response{
+	payload := jsonutils.Response{
 		Error:   false,
 		Message: "sent",
 	}
 
 	// Send the response back
-	_ = json.WriteJSON(w, http.StatusOK, payload)
+	_ = jsonutils.WriteJSON(w, http.StatusOK, payload)
 }
